refactor(study-session): drop speculative study-mode abstraction
Rename study-session/ to flashcard-session/ (session-flashcard/ to
flashcard/) and remove DeckStudyMode/study_mode: the mode-agnostic
core/flashcard composable split anticipated future study modes that
never materialized and were never branched on anywhere. session-core.ts
is renamed to session-queue.ts and reframed as a domain split (queue
scheduling vs. card-side flipping) rather than a mode boundary. Any
future study mode will be its own modal, not an extension of this
composable.
