+++ src/components/flashcard-session/composables/card-actions.ts
@@ -2,7 +2,7 @@ import { type MaybeRefOrGetter, type Ref, toValue } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useCardMutations, useCardPrompts } from '@/composables/card'
 import { useNoticeStore } from '@/stores/notice-store'
-import type { StudyCard } from './session-core'
+import type { StudyCard } from './session-queue'
 
 type UseActiveCardActionsOptions = {
   active_card: Ref<StudyCard | undefined>
