+++ src/components/flashcard-session/composables/cover-carousel.ts
