+++ src/components/flashcard-session/flashcard/index.vue
@@ -4,11 +4,11 @@ import SessionProgress from './session-progress.vue'
 import CardStage from './card-stage.vue'
 import StudyEditFooter from './study-edit-footer.vue'
 import RatingButtons from './rating-buttons/index.vue'
-import { useFlashcardSession } from '@/components/study-session/composables/flashcard-session'
-import { useCardPreview } from '@/components/study-session/composables/card-preview'
-import { useCardEdit } from '@/components/study-session/composables/card-edit'
-import { useActiveCardActions } from '@/components/study-session/composables/card-actions'
-import { useSessionCards } from '@/components/study-session/composables/session-cards'
+import { useFlashcardSession } from '@/components/flashcard-session/composables/flashcard-session'
+import { useCardPreview } from '@/components/flashcard-session/composables/card-preview'
+import { useCardEdit } from '@/components/flashcard-session/composables/card-edit'
+import { useActiveCardActions } from '@/components/flashcard-session/composables/card-actions'
+import { useSessionCards } from '@/components/flashcard-session/composables/session-cards'
 import { useModalRequestClose } from '@/composables/modal'
 import { type Grade } from 'ts-fsrs'
 import { computed, ref, useTemplateRef, watch } from 'vue'
@@ -17,7 +17,7 @@ import { useFlushDeckReviews } from '@/api/reviews'
 import { useUpsertMemberMutation } from '@/api/members'
 import { useMemberStore } from '@/stores/member'
 import { emitSfx } from '@/sfx/bus'
-import type { CardReviewResult } from '@/components/study-session/composables/session-core'
+import type { CardReviewResult } from '@/components/flashcard-session/composables/session-queue'
 
 const { decks, config_override } = defineProps<{
   decks: Deck[]
