+++ src/components/flashcard-session/flashcard/rating-buttons/advanced.vue
