+++ src/components/flashcard-session/flashcard/rating-buttons/simple.vue
