+++ src/components/flashcard-session/flashcard/session-progress.vue
