+++ src/components/flashcard-session/flashcard/study-card.vue
