+++ src/views/deck/deck-hero/study-button.vue
@@ -2,7 +2,7 @@
 import UiButton from '@/components/ui-kit/button.vue'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
-import { useStudyModal } from '@/components/study-session/composables/study-modal'
+import { useStudyModal } from '@/components/flashcard-session/composables/study-modal'
 
 type StudyButtonProps = {
   deck: Deck
